pyproject.toml
[project]
name = "gamedex"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2,<6.0",
    "pyms-django-chassis[docs,dev-tools]",
    "django-import-export>=4.0",
    "gunicorn>=21.0",
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.7.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.4",
    "pre-commit>=3.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
